Add version bump php module helper script
- Added README
- Refactored scripts
