+++ scripts/php-modules/Gemfile
@@ -5,3 +5,5 @@ ruby '~> 2.3'
 
 gem 'mini_portile', git: 'https://github.com/cf-buildpacks-eng/mini_portile'
 gem 'open-uri'
+gem 'nokogiri'
+gem 'pry'
