+++ scripts/php-modules/bump-versions.sh
@@ -5,4 +5,4 @@ set -euo pipefail
 cd "$(dirname "$0")"
 
 bundle install
-bundle exec ruby update-hashes.rb
+bundle exec ruby bump-versions.rb
