+++ tasks/build-binary-new/php73-extensions-patch.yml
@@ -1,7 +1,7 @@
 ---
 extensions:
   additions:
-    - name: gd
-      version: nil
-      md5: nil
-      klass: Gd72and73FakePeclRecipe
+  - name: gd
+    version: nil
+    md5: nil
+    klass: Gd72and73FakePeclRecipe
